+++ tests/examples/hashing/detectingEmptyParameterSet.fs
@@ -1,0 +1,12 @@
+module DetectingEmptyParameterSet
+
+open Npgsql.FSharp
+
+let transaction connection =
+    connection
+    |> Sql.connect
+    |> Sql.executeTransaction [
+        "SELECT * FROM users WHERE user_id = @user_id", [
+            [ ]
+        ]
+    ]