+++ tests/examples/hashing/errorsInTransactions.fs
@@ -8,4 +8,7 @@ let executeMultipleQueries =
     |> Sql.executeTransaction [
         "UPDATE users SET user_id = @user_id", [ ]
         "DELTE FROM meters", [ ]
+        "INSERT INTO users (username) VALUES (@username)", [
+            [ "@username", Sql.int 42 ]
+        ]
     ]