+++ tests/examples/hashing/syntaxAnalysis-usingTransactions.fs
@@ -6,7 +6,13 @@ let executeMultipleQueries =
     "connectionString"
     |> Sql.connect
     |> Sql.executeTransaction [
-        "UPDATE users SET user_id = @user_id", [ ]
+        "UPDATE users SET user_id = @user_id", [
+            [ "@user_id", Sql.int 42 ]
+            [ "@user_id", Sql.int 43 ]
+            [ "@user_id", Sql.int 44 ]
+            [ "@user_id", Sql.int 44 ]
+        ]
+
         "DELETE FROM users", [ ]
     ]